refactor: refine shared type boundaries
