+++ src/gateway/text-channel-commands.ts
@@ -8,7 +8,7 @@ import {
   mapTuiSlashCommandToGatewayArgs,
   parseTuiSlashCommand,
 } from '../tui-slash-command.js';
-import type { ArtifactMetadata } from '../types/container.js';
+import type { ArtifactMetadata } from '../types/execution.js';
 import { extractGatewayChatApprovalEvent } from './chat-approval.js';
 import {
   normalizePendingApprovalReply,
