+++ src/skills/skills-observation.ts
@@ -6,7 +6,7 @@ import {
   incrementAmendmentRunCount,
   recordSkillObservation as insertSkillObservation,
 } from '../memory/db.js';
-import type { ToolExecution } from '../types/container.js';
+import type { ToolExecution } from '../types/execution.js';
 import type {
   AdaptiveSkillsConfig,
   SkillErrorCategory,
