+++ src/types/execution.ts
@@ -1,0 +1,75 @@
+export interface RuntimeToolSchemaProperty {
+  type: string | string[];
+  description?: string;
+  items?: RuntimeToolSchemaProperty;
+  properties?: Record<string, RuntimeToolSchemaProperty>;
+  required?: string[];
+  enum?: string[];
+  minItems?: number;
+  maxItems?: number;
+}
+
+export interface RuntimeToolSchema {
+  type: 'object';
+  properties: Record<string, RuntimeToolSchemaProperty>;
+  required: string[];
+}
+
+export interface PluginRuntimeToolDefinition {
+  name: string;
+  description: string;
+  parameters: RuntimeToolSchema;
+}
+
+export interface ToolExecution {
+  name: string;
+  arguments: string;
+  result: string;
+  durationMs: number;
+  isError?: boolean;
+  blocked?: boolean;
+  blockedReason?: string;
+  approvalTier?: 'green' | 'yellow' | 'red';
+  approvalBaseTier?: 'green' | 'yellow' | 'red';
+  approvalDecision?:
+    | 'auto'
+    | 'implicit'
+    | 'approved_once'
+    | 'approved_session'
+    | 'approved_agent'
+    | 'approved_fullauto'
+    | 'promoted'
+    | 'required'
+    | 'denied';
+  approvalActionKey?: string;
+  approvalIntent?: string;
+  approvalReason?: string;
+  approvalRequestId?: string;
+  approvalExpiresAt?: number;
+  approvalAllowSession?: boolean;
+  approvalAllowAgent?: boolean;
+}
+
+export interface PendingApproval {
+  approvalId: string;
+  prompt: string;
+  intent: string;
+  reason: string;
+  allowSession: boolean;
+  allowAgent: boolean;
+  expiresAt: number | null;
+}
+
+export interface ToolProgressEvent {
+  sessionId: string;
+  toolName: string;
+  phase: 'start' | 'finish';
+  preview?: string;
+  durationMs?: number;
+}
+
+export interface ArtifactMetadata {
+  path: string;
+  filename: string;
+  mimeType: string;
+}
