+++ src/types/hybridai.ts
@@ -1,0 +1,6 @@
+export interface HybridAIBot {
+  id: string;
+  name: string;
+  description?: string;
+  model?: string;
+}
