+++ src/types/scheduler.ts
@@ -1,13 +1,3 @@
-export type ScheduleSideEffect =
-  | {
-      action: 'add';
-      cronExpr?: string;
-      runAt?: string;
-      everyMs?: number;
-      prompt: string;
-    }
-  | { action: 'remove'; taskId: number };
-
 export interface ScheduledTask {
   id: number;
   session_id: string;
@@ -21,4 +11,16 @@ export interface ScheduledTask {
   last_status: string | null;
   consecutive_errors: number;
   created_at: string;
+}
+
+// CamelCase projection of ScheduledTask passed over the container IPC boundary.
+export interface ScheduledTaskInput {
+  id: number;
+  cronExpr: string;
+  runAt: string | null;
+  everyMs: number | null;
+  prompt: string;
+  enabled: number;
+  lastRun: string | null;
+  createdAt: string;
 }
